The basic preprocessing is running.
