--- KBQA/appB/preprocessing/separated_data_files/.gitignore
@@ -1,2 +1,0 @@
-**
-!.gitignore
